migracion 1

+++ database/migrations/0001_01_01_000003_create_escuchas_table.php
@@ -11,16 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('escuchas', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('usuario_propietario_id')->constrained('users')->onDelete('cascade');
-            $table->string('nombre_asignado');
-            $table->string('telefono');
-            $table->foreignId('id_usuario_real')->nullable()->constrained('users')->onDelete('set null');
-            $table->unique(['usuario_propietario_id', 'telefono']);
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('escuchas')) {
+            Schema::create('escuchas', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('usuario_propietario_id')->constrained('users')->onDelete('cascade');
+                $table->string('nombre_asignado');
+                $table->string('telefono');
+                $table->foreignId('id_usuario_real')->nullable()->constrained('users')->onDelete('set null');
+                $table->unique(['usuario_propietario_id', 'telefono']);
+                $table->softDeletes();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
