+++ database/migrations/0001_01_01_000003_create_procesos_table.php
@@ -11,15 +11,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('procesos', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('usuario_id')->constrained('users')->onDelete('cascade');
-            $table->foreignId('escucha_id')->constrained('escuchas')->onDelete('cascade');
-            $table->string('nombre_proceso');
-            $table->timestamp('fecha_creacion')->useCurrent();
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('procesos')) {
+            Schema::create('procesos', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('usuario_id')->constrained('users')->onDelete('cascade');
+                $table->foreignId('escucha_id')->constrained('escuchas')->onDelete('cascade');
+                $table->string('nombre_proceso');
+                $table->timestamp('fecha_creacion')->useCurrent();
+                $table->softDeletes();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
